# public/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Cursos DBlandIT
  description: Una API REST que permita administrar los cursos que dicta la empresa DBlandIT.
    </br> Demo token JWT ( Todos los endpoint requieren autenticacion )
    <span style="color:#E8451C;font-weight:bold;display:block;width:600px;word-wrap:break-word;">eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.9kKTP8noR4tIDbdUEpkLHKnqGFxqAsWfIg-z4kSn5As</span>
  version: 1.1.0
servers:
  - url: http://api.iescobar.cf
    description: Live demo
  - url: http://localhost:3000
    description: Server local
tags:
  - name: "cursos"
    description: "Acceso a creacion y visualizacion de cursos"
security:
  - bearerAuth: []
paths:
  /cursos:
    get:
      summary: Retorna una lista de cursos
      tags:
        - "cursos"
      parameters:
        - in: query
          name: anno
          schema:
            type: integer
          required: false
          description: Filtrar resultados por el año indicado
        - in: query
          name: duracion
          schema:
            type: integer
          required: false
          description: Filtrar resultados por la duracion indicada
      responses:
        '200':    # status code
          description: Un array JSON que contiene la lista de cursos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Curso"
        '401':    # status code
          description: Acceso no autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Respuesta"
              example:
                msg: Acceso no autorizado
    post:
      summary: Agrega un curso a la lista
      tags:
        - "cursos"
      requestBody:
        description: Un curso para agregar a la lista
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Curso'
      responses:
        '200':    # status code
          description: Curso agregado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Curso"
              example:
                "_id": 5c415d0f155f0a6358889be5
                anno: 2020
                tema: "Bussines Inteligence"
                duracion: 30
                alumnos: []
        '401':    # status code
          description: Acceso no autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Respuesta"
              example:
                msg: Acceso no autorizado
  /cursos/{id}:
    delete:
      summary: Elimina un curso
      tags:
        - "cursos"
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id del curso que se desea eliminar
      responses:
        '200':    # status code
          description: Curso eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Curso"
              example:
                "_id": 5c415d0f155f0a6358889be5
                anno: 2020
                tema: "Bussines Inteligence"
                duracion: 30
                alumnos: []

        '404':    # status code
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Respuesta"
              example:
                msg: "No se pudo eliminar el curso: [Descripcion del error]"
        '401':    # status code
          description: Acceso no autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Respuesta"
              example:
                msg: Acceso no autorizado
  /cursos/{id}/alumnos:
    get:
      summary: Retorna una lista de alumnos
      tags:
        - "cursos"
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id del curso del cual se desean consultar los alumnos
        - in: query
          name: destacado
          schema:
            type: boolean
          required: false
          description: (True) Obtener el alumno con nota mas alta
      responses:
        '200':    # status code
          description: Un array JSON de alumnos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alumno"
        '404':    # status code
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Respuesta"
              example:
                msg: "No se pudieron recuperar los alumnos: [Descripcion del error]"
        '401':    # status code
          description: Acceso no autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Respuesta"
              example:
                msg: Acceso no autorizado

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Alumno:
      type: object
      properties:
        apellido:
          type: "string"
        nombre:
          type: "string"
        direccion:
          type: "string"
        dni:
          type: "integer"
        nota:
          type: "integer"
      example:
        dni: 37161487
        nombre: Brewer
        apellido: Freeman
        nota: 7
    Curso:
      type: object
      properties:
        "_id":
          type: "string"
        anno:
          type: "integer"
        tema:
          type: "string"
        duracion:
          type: "integer"
        alumnos:
          type: "array"
          items:
            $ref: "#/components/schemas/Alumno"
      example:
        anno: 2020
        tema: "Bussines Inteligence"
        duracion: 30
        alumnos: []
    Respuesta:
      type: object
      properties:
        msg:
          type: "string"